[New] External Tools: Launch system tools, like terminal or file manager
This plugin serves to quickly launch external tools in the context
of the current document, i.e. in the same directory, or with focus
on the file. The current implementation includes a Terminal and a
File Manager.

The commands to execute are configurable in the Plugin preferences.
The tools may be launched from the Tools menu, or via keybinds.

The code is modular to support easy addition of extra tools, though
in the feature it would be really nice to support such additions
directly in the preferences without code modification.