codenav/goto_file: Keep dialog open after denying creation of not-found file
Previously, if the file wasn't found and you replied 'no' to the prompt
for creating it, all dialogs closed. This can be annoying in situations
where the name/path was entered wrong, and you have to start all over
again after receiving the file not found error. This commit allows one
to amend the mistake from right where they left off.

It was deemed necessary to add GTK_DIALOG_MODAL flag to the dialog creation.
Without it, after the repeat gtk_dialog_run call (through the goto), the UI
got kinda bugged, with the focus (correctly) being in the dialog, but visually
appearing as if it had shifted back to the editor pane.