codenav/goto_file: Rework directory_check; simpler, better, safer
The new directory_check is (IMHO) simpler, easier to follow and maintain,
and less prone to edge cases.

As part of simplifications, do away with the "reference model". We do keep
track of the "reference dir" and if need be simply calculate a mode for it
as if any other dir.

Bonus (admittedly rare) issue solved: not having to close and reopen the
dialog if a file is created while it's opened...